api endpoit logout, register and list of users

+++ app/Http/Controllers/api/Auth/AuthenticatedTokenController.php
@@ -4,23 +4,13 @@ namespace App\Http\Controllers\Api\Auth;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Auth\LoginRequest;
-use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
-use Illuminate\View\View;
 use Exception;
 use Illuminate\Http\JsonResponse;
 
 class AuthenticatedTokenController extends Controller
 {
-    /**
-     * Display the login view.
-     */
-    public function create(): View
-    {
-        return view('auth.login');
-    }
-
     /**
      * Handle an incoming authentication request.
      */
@@ -48,15 +38,23 @@ class AuthenticatedTokenController extends Controller
         }
     }
 
-    /**
-     * Destroy an authenticated session.
-     */
-    public function destroy(Request $request)
+    public function destroy(Request $request): JsonResponse
     {
-        $request->user()->tokens()->delete();
-        Auth::guard('web')->logout();
-        $request->session()->invalidate();
-        $request->session()->regenerateToken();
-        //return redirect('login');
+        try {
+
+            if ($request->user()) {
+                $request->user()->tokens()->delete();
+            }
+
+            Auth::guard('web')->logout();
+
+            return response()->json([
+                'message' => 'Sesión cerrada correctamente'
+            ], 200);
+        } catch (Exception $error) {
+            return response()->json([
+                'error' => 'Error al cerrar sesión: '.$error->getMessage()
+            ], 500);
+        }
     }
 }
